Moved makeRole function to a functions.js file
